+++ src/main/java/com/github/hcsp/Crawler.java
@@ -1,0 +1,104 @@
+package com.github.hcsp;
+
+import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
+import org.apache.http.HttpEntity;
+import org.apache.http.client.methods.CloseableHttpResponse;
+import org.apache.http.client.methods.HttpGet;
+import org.apache.http.impl.client.CloseableHttpClient;
+import org.apache.http.impl.client.HttpClients;
+import org.apache.http.util.EntityUtils;
+import org.jsoup.Jsoup;
+import org.jsoup.nodes.Document;
+import org.jsoup.nodes.Element;
+
+import java.io.IOException;
+import java.sql.*;
+import java.util.ArrayList;
+import java.util.stream.Collectors;
+
+public class Crawler {
+    private CrawlerDao dao = new JdbcCrawlerDao();
+
+    public void run() throws SQLException, IOException {
+        String link;
+
+        while ((link = dao.getNextLinkThenDelete()) != null) {
+            //先从数据库里拿出来一个链接，并从数据库中删除掉，准备处理之
+            if (dao.isLinkedProcessed(link)) {
+                continue;
+            }
+
+            if (isInterestingLink(link)) {
+                //这是我们感兴趣的，我们只处理新浪站内的链接
+                System.out.println(link);
+                Document doc = httpGetAndParseHtml(link);
+                parseUrlsFromPageAndStoreIntoDatabase(doc);
+                storeIntoDatabaseIfItIsNewPage(doc, link);
+                dao.updateDatabase(link, "INSERT INTO LINKS_ALREADY_PROCESSED (link) values(?) ");
+            }
+        }
+
+    }
+
+    @SuppressFBWarnings("DMI_CONSTANT_DB_PASSWORD")
+    public static void  main(String[] args) throws IOException, SQLException {
+        new Crawler().run();
+    }
+
+    private void parseUrlsFromPageAndStoreIntoDatabase(Document doc) throws SQLException {
+        for (Element aTag : doc.select("a")) {
+            String href = aTag.attr("href");
+
+            if (href.startsWith("//")) {
+                href = "https:" + href;
+            }
+
+            if (!href.toLowerCase().startsWith("javascript")) {
+                dao.updateDatabase(href, "INSERT INTO LINKS_TO_BE_PROCESSED(link) values(?)");
+            }
+        }
+    }
+
+
+    private static Document httpGetAndParseHtml(String link) throws IOException {
+        CloseableHttpClient httpclient = HttpClients.createDefault();
+        HttpGet httpGet = new HttpGet(link);
+        httpGet.addHeader("User-agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_1_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.96 Safari/537.36");
+
+        try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
+            HttpEntity entity1 = response1.getEntity();
+            String html = EntityUtils.toString(entity1);
+            return Jsoup.parse(html);
+        }
+    }
+
+    private static boolean isInterestingLink(String link) {
+        return (isNewsPage(link) || isIndexPage(link) && isNotLoginPage(link));
+    }
+
+    private static boolean isNotLoginPage(String link) {
+        return !link.contains("passport.sina.cn");
+    }
+
+    private static boolean isIndexPage(String link) {
+        return "https://sina.cn".equals(link);
+    }
+
+
+    private static boolean isNewsPage(String link) {
+        return link.contains("news.sina.cn");
+    }
+
+
+    private void storeIntoDatabaseIfItIsNewPage(Document doc, String link) throws SQLException {
+        ArrayList<Element> articleTags = doc.select("article");
+        if (!articleTags.isEmpty()) {
+            for (Element articleTag : articleTags) {
+                String title = articleTags.get(0).child(0).text();
+                String content = articleTag.select("p").stream().map(Element::text).collect(Collectors.joining("\n"));
+                dao.insertNewsIntoDatabase(link, title, content);
+
+            }
+        }
+    }
+}
