--- src/main/java/com/github/hcsp/Main.java
@@ -1,154 +1,0 @@
-package com.github.hcsp;
-
-import edu.umd.cs.findbugs.annotations.SuppressFBWarnings;
-import org.apache.http.HttpEntity;
-import org.apache.http.client.methods.CloseableHttpResponse;
-import org.apache.http.client.methods.HttpGet;
-import org.apache.http.impl.client.CloseableHttpClient;
-import org.apache.http.impl.client.HttpClients;
-import org.apache.http.util.EntityUtils;
-import org.jsoup.Jsoup;
-import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-
-import java.io.IOException;
-import java.sql.*;
-import java.util.ArrayList;
-import java.util.stream.Collectors;
-
-public class Main {
-    private static final String USER_NAME = "root";
-    private static final String PASSWORD = "root";
-
-
-    private static String getNextLink(Connection connection, String sql) throws SQLException {
-        ResultSet resultSet = null;
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            resultSet = statement.executeQuery();
-            while (resultSet.next()) {
-                return resultSet.getString(1);
-            }
-        } finally {
-            if (resultSet != null) {
-                resultSet.close();
-            }
-        }
-        return null;
-    }
-
-    private static String getNextLinkThenDelete(Connection connection) throws SQLException {
-        String link = getNextLink(connection, "select  link from LINKS_TO_BE_PROCESSED limit 1");
-
-        if (link != null) {
-            updateDatabase(connection, link, "delete  from LINKS_TO_BE_PROCESSED where link = ?");
-        }
-        return link;
-    }
-
-    @SuppressFBWarnings("DMI_CONSTANT_DB_PASSWORD")
-    public static void main(String[] args) throws IOException, SQLException {
-        Connection connection = DriverManager.getConnection("jdbc:h2:file:/Users/niuweizhe/mycrawler/news", USER_NAME, PASSWORD);
-
-        String link;
-
-        while ((link = getNextLinkThenDelete(connection)) != null) {
-            //先从数据库里拿出来一个链接，并从数据库中删除掉，准备处理之
-            if (isLinkedProcessed(connection, link)) {
-                continue;
-            }
-
-            if (isInterestingLink(link)) {
-                //这是我们感兴趣的，我们只处理新浪站内的链接
-                System.out.println(link);
-                Document doc = httpGetAndParseHtml(link);
-                parseUrlsFromPageAndStoreIntoDatabase(connection, doc);
-                storeIntoDatabaseIfItIsNewPage(connection, doc, link);
-                updateDatabase(connection, link, "INSERT INTO LINKS_ALREADY_PROCESSED (link) values(?) ");
-            }
-        }
-    }
-
-    private static void parseUrlsFromPageAndStoreIntoDatabase(Connection connection, Document doc) throws SQLException {
-        for (Element aTag : doc.select("a")) {
-            String href = aTag.attr("href");
-
-            if (href.startsWith("//")) {
-                href = "https:" + href;
-            }
-
-            if (!href.toLowerCase().startsWith("javascript")) {
-                updateDatabase(connection, href, "INSERT INTO LINKS_TO_BE_PROCESSED(link) values(?)");
-            }
-        }
-    }
-
-    private static boolean isLinkedProcessed(Connection connection, String link) throws SQLException {
-        ResultSet resultSet = null;
-        try (PreparedStatement statement = connection.prepareStatement("select LINK from LINKS_ALREADY_PROCESSED where link = ? ")) {
-            statement.setString(1, link);
-            resultSet = statement.executeQuery();
-            while (resultSet.next()) {
-                return true;
-            }
-        } finally {
-            if (resultSet != null) {
-                resultSet.close();
-            }
-        }
-        return false;
-    }
-
-    private static void updateDatabase(Connection connection, String link, String sql) throws SQLException {
-        try (PreparedStatement statement = connection.prepareStatement(sql)) {
-            statement.setString(1, link);
-            statement.executeUpdate();
-        }
-    }
-
-    private static Document httpGetAndParseHtml(String link) throws IOException {
-        CloseableHttpClient httpclient = HttpClients.createDefault();
-        HttpGet httpGet = new HttpGet(link);
-        httpGet.addHeader("User-agent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 11_1_0) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/88.0.4324.96 Safari/537.36");
-
-        try (CloseableHttpResponse response1 = httpclient.execute(httpGet)) {
-            HttpEntity entity1 = response1.getEntity();
-            String html = EntityUtils.toString(entity1);
-            return Jsoup.parse(html);
-        }
-    }
-
-    private static boolean isInterestingLink(String link) {
-        return (isNewsPage(link) || isIndexPage(link) && isNotLoginPage(link));
-    }
-
-    private static boolean isNotLoginPage(String link) {
-        return !link.contains("passport.sina.cn");
-    }
-
-    private static boolean isIndexPage(String link) {
-        return "https://sina.cn".equals(link);
-    }
-
-
-    private static boolean isNewsPage(String link) {
-        return link.contains("news.sina.cn");
-    }
-
-
-    private static void storeIntoDatabaseIfItIsNewPage(Connection connection, Document doc, String link) throws SQLException {
-        ArrayList<Element> articleTags = doc.select("article");
-        if (!articleTags.isEmpty()) {
-            for (Element articleTag : articleTags) {
-                String title = articleTags.get(0).child(0).text();
-                String content = articleTag.select("p").stream().map(Element::text).collect(Collectors.joining("\n"));
-
-                try (PreparedStatement statement = connection.prepareStatement("insert into NEWS(url,title,content,CREATED_AT,MODIFIED_AT) values ( ?,?,?,now(),now())")) {
-                    statement.setString(1, link);
-                    statement.setString(2, title);
-                    statement.setString(3, content);
-                    statement.executeUpdate();
-                }
-            }
-        }
-    }
-}
